pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mini-azurestorage-mcp"
version = "0.1.0"
description = "A stdio MCP server that uploads local files to Azure Blob Storage"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.27.0",
    "azure-storage-blob>=12.28.0",
]

[project.scripts]
mini-azurestorage-mcp = "server:main"

[tool.hatch.build.targets.wheel]
packages = ["."]
